move reducers to common base src dir
